Moved Data Access into actor
